backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pymongo==4.6.1
motor==3.3.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.11.3
email-validator==2.1.0
dnspython==2.4.2
